feat: initial implementation of dsdb_hw1
This commit will provide a first implementation of the homework request. It is a user management service that exposes REST and
gRPC interfaces, and a data collector that retrieves flight information from the OpenSky APIs based on user interests.
Ths entire setup is orchestrated via `docker-compose` with dedicated PostgreSQL databases.